Add Objective-C API
