+++ Package.swift
@@ -221,7 +221,9 @@ let package = Package(
                 "Realm/RLMUpdateResult.mm",
                 "Realm/RLMUser.mm",
                 "Realm/RLMUserAPIKey.mm",
-                "Realm/RLMFlexibleSyncConfiguration.mm"
+                "Realm/RLMFlexibleSyncConfiguration.mm",
+                "Realm/RLMSubscriptionSet.mm",
+                "Realm/RLMSubscription.mm"
             ],
             publicHeadersPath: "include",
             cxxSettings: cxxSettings
