+++ Realm/RLMSubscription.mm
@@ -16,16 +16,26 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-#import "RLMFlexibleSyncConfiguration_Private.hpp"
+#import "RLMSubscription.h"
 
-#import "RLMUser_Private.hpp"
+// TODO: Flexible Sync - Add docstrings
 
-@implementation RLMFlexibleSyncConfiguration
+@implementation RLMSubscriptionTask
 
-- (instancetype)initWithUser:(RLMUser *)user {
+@end
+
+@implementation RLMSubscription
+
+- (instancetype)initWithName:(nullable NSString *)name
+                   predicate:(NSPredicate *)predicate {
     // TODO: Flexible Sync - Add initialiser implementation
     [NSException raise:@"NotImplemented" format:@"Needs Impmentation"];
     return NULL;
+}
+
+- (void)updateSubscriptionWithPredicate:(NSPredicate *)predicate
+                                  error:(NSError **)error {
+    [NSException raise:@"NotImplemented" format:@"Needs Impmentation"];
 }
 
 @end
